feat(terraform): add DNS managed zone configuration
Add a new file `dns_managed_zone.tf` to configure a Google Cloud DNS managed zone. This managed zone is named "zone-1" and has the DNS name "nginx.fajarmaftuhfadli.com.". It is associated with the main project in the Google Cloud project.

feat(terraform): add A record to DNS managed zone

Add a new resource `google_dns_record_set.a` to configure an A record in the DNS managed zone. This A record has a TTL of 300 seconds and points to the NAT IP of the `google_compute_instance.instance-1` resource.

feat(terraform): comment out CNAME record configuration

Comment out the configuration for the CNAME record in the DNS managed zone.

feat(terraform): comment out GKE IAM member configuration

Comment out the configuration for the IAM members in the GKE project.